Fix a simple format string argument mismatch.
  CC       cckddasd.lo
In file included from hercules.h:104,
                 from cckddasd.c:21:
cckddasd.c: In function 'cckd_trace':
cckddasd.c:6715:13: warning: format '%ld' expects argument of type 'long int', but argument 5 has type 'suseconds_t' {aka 'int'} [-Wformat=]
 6715 |             "%s.%6.6ld %1d:%04X ",          // "hh:mm:ss.uuuuuu n:CCUU "
      |             ^~~~~~~~~~~~~~~~~~~~~
 6716 |             todwrk + 11,                    // "hh:mm:ss" (%s)
 6717 |             timeval.tv_usec,                // "uuuuuu"   (%6.6ld
      |             ~~~~~~~~~~~~~~~
      |                    |
      |                    suseconds_t {aka int}
msgenu.h:208:70: note: in definition of macro 'MSGBUF'
  208 | #define MSGBUF( _buf, ... )     snprintf(_buf, sizeof(_buf),      ## __VA_ARGS__ )
      |                                                                      ^~~~~~~~~~~
cckddasd.c:6715:22: note: format string is defined here
 6715 |             "%s.%6.6ld %1d:%04X ",          // "hh:mm:ss.uuuuuu n:CCUU "
      |                 ~~~~~^
      |                      |
      |                      long int
      |                 %6.6d

  CC       tcpnje.o
tcpnje.c: In function 'tcpnje_thread':
tcpnje.c:1848:25: warning: format '%ld' expects argument of type 'long int', but argument 4 has type 'suseconds_t' {aka 'int'} [-Wformat=]
 1848 |             DBGMSG(512, "HHCTN127D %4.4X:TCPNJE - select() timeout after %ld seconds %ld microseconds\n",
      |                         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 1849 |                         devnum, tvcopy.tv_sec, tvcopy.tv_usec);
      |                                                ~~~~~~~~~~~~~~
      |                                                      |
      |                                                      suseconds_t {aka int}
tcpnje.c:98:109: note: in definition of macro 'DBGMSG'
   98 | #define DBGMSG(_level, ...) if ((tn->dev->ccwtrace && (_level & tn->trace)) || (_level & tn->debug)) logmsg(__VA_ARGS__)
      |                                                                                                             ^~~~~~~~~~~
tcpnje.c:1848:88: note: format string is defined here
 1848 |             DBGMSG(512, "HHCTN127D %4.4X:TCPNJE - select() timeout after %ld seconds %ld microseconds\n",
      |                                                                                      ~~^
      |                                                                                        |
      |                                                                                        long int
      |                                                                                      %d

Since the argument is apparently too short, it is easy to convert it to
a longer type. On some systems it may already be long int, so changing
the format string to %d is likely to be incorrect in such cases.

+++ tcpnje.c
@@ -1846,7 +1846,7 @@ static void *tcpnje_thread(void *vtn)
         if (selectcount == 0)
         {
             DBGMSG(512, "HHCTN127D %4.4X:TCPNJE - select() timeout after %ld seconds %ld microseconds\n",
-                        devnum, tvcopy.tv_sec, tvcopy.tv_usec);
+                        devnum, tvcopy.tv_sec, (long int)tvcopy.tv_usec);
 
             /* Reset Call issued flag */
             tn->callissued = 0;
